+++ src/main/java/com/guxingke/intellij/plugin/postfix/ToolsPostfixTemplateProvider.java
@@ -2,9 +2,11 @@ package com.guxingke.intellij.plugin.postfix;
 
 import com.guxingke.intellij.plugin.Configs;
 import com.guxingke.intellij.plugin.postfix.template.TemplateFactory;
+import com.guxingke.intellij.plugin.postfix.template.struct.DynamicStructMapperPostfixTemplate;
 import com.guxingke.intellij.plugin.postfix.template.struct.StructMapperPostfixTemplate;
 import com.intellij.codeInsight.template.postfix.templates.PostfixTemplate;
 import com.intellij.codeInsight.template.postfix.templates.PostfixTemplateProvider;
+import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.editor.Editor;
 import com.intellij.psi.PsiFile;
 import java.util.Collection;
@@ -16,19 +18,35 @@ import org.jetbrains.annotations.NotNull;
 
 public class ToolsPostfixTemplateProvider implements PostfixTemplateProvider {
 
-  private final Set<PostfixTemplate> templates;
+  private static final Logger log = Logger.getInstance(ToolsPostfixTemplateProvider.class);
+
+  private long lastGet = System.currentTimeMillis();
+
+  private Set<PostfixTemplate> templates;
+  private boolean init = false;
 
   public ToolsPostfixTemplateProvider() {
+    this.templates = loadTemplates();
+  }
+
+  private synchronized Set<PostfixTemplate> loadTemplates() {
+    log.info("load templates begin ...");
     var cfg = Configs.getConfig();
+    log.info("config mode debug: " + cfg.isDebug());
+    if (init && !cfg.isDebug()) { // return exists
+      log.info("reuse exists templates " + templates.size());
+      return templates;
+    }
+
     if (!cfg.getPostfix().isEnable()) {
-      templates = new HashSet<>();
-      return;
+      log.info("postfix status is global disable");
+      return new HashSet<>();
     }
 
     var ic = cfg.getPostfix().getInternal();
     Set<PostfixTemplate> internals = new HashSet<>();
     if (ic.isEnable()) {
-      internals = Stream.of(new StructMapperPostfixTemplate(this))
+      internals = Stream.of(new StructMapperPostfixTemplate(this), new DynamicStructMapperPostfixTemplate(this))
           .filter(it -> !ic.getBlocklist().contains(it.getPresentableName()))
           .collect(Collectors.toSet());
     }
@@ -51,14 +69,23 @@ public class ToolsPostfixTemplateProvider implements PostfixTemplateProvider {
           .collect(Collectors.toSet());
     }
     // 按触发词去重, external -> builtin -> internal 优先级
-    templates = new HashSet<>(Stream.of(externals, builtinTemplates, internals)
-                                  .flatMap(Collection::stream)
-                                  .collect(Collectors.toMap(PostfixTemplate::getPresentableName, it -> it, (l, r) -> r))
-                                  .values());
+    init = true;
+    var ts = Stream.of(externals, builtinTemplates, internals)
+        .flatMap(Collection::stream)
+        .collect(Collectors.toMap(PostfixTemplate::getPresentableName, it -> it, (l, r) -> r))
+        .values();
+    log.info("load templates done " + ts.size());
+    return new HashSet<>(ts);
   }
 
   @Override
   public @NotNull Set<PostfixTemplate> getTemplates() {
+    var now = System.currentTimeMillis();
+    if (now - lastGet > 60 * 1000) { // 1 min
+      this.templates = loadTemplates();
+      lastGet = now;
+    }
+    log.info("got templates " + templates.size());
     return templates;
   }
 
@@ -66,6 +93,7 @@ public class ToolsPostfixTemplateProvider implements PostfixTemplateProvider {
   public boolean isTerminalSymbol(char currentChar) {
     return currentChar == '.';
   }
+
   @Override
   public void preExpand(
       @NotNull PsiFile file,