+++ CHANGELOG.md
@@ -7,7 +7,9 @@
 ### Added
 
 - struct mapper type converter.
- 
+- struct mapper rule config.
+- debug mode for debugging config, be reload user config at most 60 seconds.
+
 ## [0.0.5] - 2022-10-09
 
 ### Added